third_party/WebKit/LayoutTests/fast/css/webkit-text-display-none-expected.txt

This test checks that setting display to none on the body element does not change -webkit-text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(d).getPropertyValue('color') is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
